requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
pillow==10.4.0
ratelimit==2.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
